ujcuccot

+++ m-indep.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 function m.src_indep(){
-    expsource ${WGPATH}/mtool/m-indep.sh;
+    source ${WGPATH}/mtool/m-indep.sh;
 }
 
 function btofslash () { echo "$1" | sed 's|\\|/|g'; }
@@ -12,7 +12,6 @@ function m.amon() {
     touch "${al}";
     tail -n 0 -f ${al};
 }
-
 
 function m.tonice(){
     echo "$1" | sed 's/\ \ */\ /g' | sed 's/\ $//' | sed 's/\ /\_/g' | sed 's/\./\-/g';
@@ -44,14 +43,31 @@ function m.countchar(){
 
 function m.isnumber(){
     if [[ ! "$1" =~ ^(0|[-]?[1-9]+[0-9]*)$ ]]; then
-         
-         
         return 1;  
     else
          
         return 0;
     fi
 }
+
+
+# handles lofasz.exr and lofasz.3.exr -> lofasz
+function m.fn_base() {
+   local fba="${1%.*}";
+   echo "${fba%.*}";
+}
+
+# color.1.exr tibor.exr -> exr
+function m.fn_ext(){
+   echo ".${1#*.}";
+}
+
+# color.1.exr -> exr
+function m.fn_mext() {
+   local fba="${1#*.}";
+   echo "${fba#*.}";
+}
+
 
 # <root> <depth>
 function m.dtree() {
@@ -76,50 +92,158 @@ function m.runftree(){
     
 }
 
-
+# the last .ext only
 function m.repext(){
     if [[ -z "$1" ]]; then
         echo "***ERROR no input path!"
         return 0;
     fi
-    local outf="${1%.*}";
-    echo "${outf}.$2";
+    
+    local mext=$(m.fn_mext "${1}");
+    echo "${1}" | sed "s|${mext}|${2}|g";
+   
 }
 
-function m.dupe(){
-    if [[ -z "$1" ]]; then
-        echo "***ERROR no input path!"
-        return 0;
-    fi
-    local outf="${1%.*}";
-    echo "${outf}_.$2";
-}
 
 function m.tmpfile(){
     if [[ -z "$1" ]]; then
         echo "***ERROR no input path!"
         return 0;
     fi
-    local outf="${1%.*}_";
-    local inext="${k#*.}";
+    local fb=$(m.fn_base "${1}");
+    local outf="${fb}_";
+    local inext=$(m.fn_ext "${1}");
     local outext="${inext}";
     if [[ -n "$2" ]]; then
-        outext="$2";
+        outext="${2}";
     fi
-    echo "${outf}.${outext}"
+    echo "${outf}${outext}"
 }
 
 
+function m.side(){
+    if [[ -z "$1" ]]; then
+        echo "***ERROR no input path!"
+        return 0;
+    fi
 
-function m.title() {
-  if [[ -z "$ORIG" ]]; then
-    ORIG=$PS1
-  fi
-  TITLE="\[\e]2;$*\a\]"
-  PS1=${ORIG}${TITLE}
+    postfix='';
+    if [[ ( -n "${3}" ) ]]; then
+        postfix="${3}";
+    fi
+
+    local fb=$(m.fn_base "${1}");
+    local outf="${fb}${postfix}";
+    local origext=$(m.fn_ext "${1}");
+    local outext="${origext}";
+
+
+    local res="${outf}${outext}";
+
+    if [[ ( -n "$2" ) && ( "${2}" != '_' ) ]]; then
+        res=$(m.repext "${res}" "${2}");
+    fi
+
+    echo "${res}";
 }
 
 
+function m.namekey(){
+    local fb=$(basename "${1}");
+    echo "${2}${fb}" | sed 's|\.||g' ;
+}
+
+function m.waituntilany() {
+    local wdir="${1}";
+    local wpat="${2}";
+    sleep 2;
+    local nofbusyfiles=$(find "${wdir}" -name "${wpat}" -type f | wc -l);
+    while [[ ${nofbusyfiles} -gt 0 ]]
+    do
+        sleep 2;
+        nofbusyfiles=$(find "${wdir}" -name "${wpat}" -type f | wc -l);
+    done
+}
+
+# dir nof
+function m.waitfornof() {
+    local wdir="${1}";
+    local wnof=${2};
+    local nof=$(ls -1 "${wdir}" | wc -l);
+    while [[ ! ${nof} -eq ${wnof} ]]
+    do
+        sleep 1;
+        nof=$(ls -1 "${wdir}" | wc -l);
+    done;
+}
+
+# remove all new line 
+function m.singleline() {
+    echo "$(sed ':a;N;$!ba;s/\n/,/g' ${1})";
+}
+
+# max 10 char long random string 
+function m.rndstr(){
+    local yearday=`date +%j`;
+    local linsec=`date +%s`;
+    local ns=`date +%N`;
+    ns=${ns%%00};
+    sleep 0.1;
+    local nns=`date +%N`;
+    nns=${nns%%00};
+    echo "${yearday}${linsec:(-3)}${ns:0:2}${nns:0:2}";
+}
+
+# remove white space
+function m.remws(){
+    echo ${1} | sed 's|\s||g';
+}
+
+# dir patFile ext
+function m.noffiles(){
+    #echo "ls -1 "${1}/"*"${2}"*"${3}" | wc -l;";
+    local lista=$(ls -1d "${1}/"*"${2}"*"${3}" 2>/dev/null) ;
+    if [[ $? -ne 0 ]]; then
+        echo '0';
+    fi
+ 
+    local szavak=$(echo "${lista}" | wc -w);
+    if [[ ${szavak} -eq 0 ]]; then
+        # nincsenek szavak - ures sor / sorok
+        echo '0';
+    else
+        echo "${lista}" | wc -l ;
+    fi
+
+}
+
+# sec -> Xmin Ysec
+function m.sec2min(){
+    local em=$((${1} / 60));
+    local rs=$((${1} % 60));
+    if [[ ( -z "${em}" ) || ( ${em} -eq 0 )]]; then
+        echo "${1}s";
+    else
+        echo "${em}min ${rs}s";
+    fi
+}
+
+
+# remove ws 
+function m.rwspace(){
+    echo "${1}" | sed -e 's/\s\+//g';
+}
+
+function m.allchannels(){
+    local srci="${1}";
+    local ach=$(oiiotool.exe --info -v "${srci}" | grep -e "channel list:" | sed 's|channel list:||g' );
+    echo "${ach}";
+}
+
+function m.allch_except(){
+    local aexc=$(echo $(m.allchannels "${1}") | sed 's|'"${2}"'||g' );
+    m.rwspace "${aexc}";
+}
 
 
 # ----------- up: truely independent, down: dependent only from above -----
@@ -127,7 +251,7 @@ function m.title() {
 
 
 m.txnext(){
-    local res=$(m.repext $1 'tx');
+    local res=$(m.repext "$1" 'tx');
     echo "${res}";
 }
 
@@ -136,30 +260,38 @@ m.float2graytx() {
     local pin="$1";
      local tx=$(m.txnext "${pin}");
     local gray=$(m.tmpfile "${pin}" 'exr');
-    oiiotool.exe -i "${pin}" --chsum:weight=.2126,.7152,.0722 --ch 0,0,0 -o ${gray}
-    maketx --hicomp --oiio --fixnan box3 -d float -u --stats --filter radial-lanczos3 --format exr -o "${tx}"  "${gray}"  >/dev/null 2>&1;
+    oiiotool.exe -i "${pin}" --iscolorspace linear --chsum:weight=.2126,.7152,.0722 --ch 0,0,0 -o ${gray}
+    maketx --hicomp --oiio --fixnan box3 -d float -u --stats --filter radial-lanczos3 --format exr -o "${tx}"  "${gray}"  >/dev/null;
     rm -f "${gray}";
     echo "${tx}";
 }
 
 m.float2tx() {
     local tx=$(m.txnext "${1}");
-    maketx --hicomp --oiio --fixnan box3 -d float -u --stats --filter radial-lanczos3 --format exr -o "${tx}" "$1"  >/dev/null 2>&1;
+    maketx --hicomp --oiio --fixnan box3 -d float -u --stats --filter radial-lanczos3 --format exr -o "${tx}" "$1"  >/dev/null;
     echo "${tx}";
 }
 
 
 m.8bit2tx(){
     local pin="$1";
-    local lin=$(m.tmpfile "${pin}" 'png');
-    local tx=$(m.txnext "${pin}");
-    # any sRGB -> 8bit linear png 
-    oiiotool.exe -i "${pin}" --colorconvert sRGB linear -o "${lin}"
-    maketx --oiio --fixnan box3 -d uint8 -u --stats --filter radial-lanczos3 --format tiff -o "${tx}" "${lin}"  >/dev/null 2>&1;
+    if [[ ! -e "${1}" ]]; then
+        echo "file not found:[${pin}]";
+        exit 1;
+    fi
+    
+    pin=$(realpath "${pin}");
+
+    # any sRGB -> 8bit linear png
+    local lin=$(m.side "${pin}"  'png' '_linear' );
+    local tx=$(m.side "${pin}" 'tx' '' );
+    #echo "lin:[${lin}]";
+    #echo "tx: [${tx}";
+    oiiotool.exe -i "${pin}" --colorconvert sRGB linear -o "${lin}"  >/dev/null;
+    maketx --oiio -d uint8 -u --format tiff -o "${tx}" "${lin}"  >/dev/null;
     rm -f "${lin}";
     echo "${tx}";
 }
-
 
 
 #  1       2       3      
@@ -230,122 +362,6 @@ m.vfxext() {
     echo "${res}";
 }
 
-function m.fstore() { 
-    filedata[${#filedata[@]}]=$1
-}
-
-
-# 0 - OK - do nothing
-# 1 - empty path - nonfixable
-# 2 - no extension - nonfixable
-# 3 - fixable - fix it!
-m.dfile() {
-    
-    local isnic=0; 
-    local n="${1}";
-    local fnoext="${n%.*}"; # no ext
-    local ffile="${n##*/}"; # full filename
-    local fdir="${n%/*}"; # directory part
-    local parent=${fdir##*/} # parent
-    local vext=$(m.vfxext "${ffile}");
-    local vname=${ffile%"${vext}"};
-    
-    # cannot handle case
-    if [[ ${#n} == 0 ]]; then
-        #echo "ERR wtf, empty file to check for nice...";
-        return 1; # don't work with empty path
-    fi
-
-    # cannot handle case
-    # zero dots?
-    local fdot=$(m.dotcount "{$ffile}");
-    if [[ $fdot == 0 ]]; then
-        #echo "SKIP file with no extension ${n}"
-        return 2; # well valid, don't touch it!
-    fi
-
-     
-    # dots > 2 ?
-    local dots=$(m.dotcount "${vname}");
-    if [[ $dots > 0 ]]; then
-        isnic=3;
-    fi
-
-    # white space?
-    local hasws=0;
-    m.hasws "${vname}";
-    if [[ $? == 1 ]]; then
-        isnic=3;
-        hasws=1;
-    fi
-
-    MVRB=1;
-    if [[ $MVRB == 1 ]]; then
-        echo "        orig ${n}";
-        echo "         dir ${fdir}";
-        echo "      no ext ${fnoext}";
-        echo "      parent ${parent}";
-        echo "   full file ${ffile}";
-        echo "    vfx name ${vname}";    
-        echo "     vfx ext ${vext}";
-        echo "     is nice ${isnic}";
-        echo "      has ws ${hasws}";
-        echo "        dots ${dots}";
-    fi
-
-    m.fstore $fdir;
-    m.fstore $vname;
-    m.fstore $vext;
-
-    return $isnic;
-
-}
-
-  
-
- 
-m.isfixable(){
-    
-
-    m.dfile "${1}";
-    local res=$?;
-    echo "dfile res:${res}----${1}";
-    if [[ $res == 3 ]]; then
-         
-        return 0; #fixable
-    fi
-    
-    if [[ $res == 0 ]]; then
-         
-        return 1; #ok
-    else 
-        
-        return 1; #unfixable 1 and 2
-    fi
-}
-
-
-
-m.fix2nice(){
-    declare -a -g filedata;
-    local fn="${1}";
-    #echo "${fn}";
-    local np="";
-    m.isfixable "${fn}";
-    #echo "${fn}";
-    local res=$?;
-    #echo "[res:${res}][${filedata[0]}][${filedata[1]}][${filedata[2]}]";
-    local vname="${filedata[1]}";
-    echo "vn-(${vname})";
-    if [[ ${res} == 0 ]]; then
-         echo "fixable ${fn}";
-         vname=${filedata[1]};
-         np=$(m.tonice "${vname}");
-         echo "javitott[${np}]";
-         echo ' '
-    fi
-    unset -v filedata;
-}
 
  
 m.linpath(){
@@ -412,7 +428,7 @@ m.winpath(){
 }
  
 
-export -f m.fix2nice  m.tonice   m.dfile m.vfxext m.countchar m.hasws m.isseq m.seqext m.dotcount m.isfixable m.fstore 
+export -f m.rwspace m.allchannels m.allch_except m.side  m.fn_base m.fn_ext m.fn_mext m.namekey m.waituntilany m.singleline m.rndstr m.remws  m.waitfornof m.noffiles m.sec2min;
  
 
 # -----------------------------------------
@@ -427,11 +443,11 @@ mtool=$(m.linpath "${WGPATH}/mtool");
 echo "mtool root: [${mtool}]";
 
 # temp files
-[ ! -e "${mtool}/_tmp" ] && mkdir -p "${mtool}/_tmp";
-export MTRASH="${mtool}/_tmp";
+#[ ! -e "${mtool}/_tmp" ] && mkdir -p "${mtool}/_tmp";
+#export MTRASH="${mtool}/_tmp";
 
 # fusion comp templates
-export MTEMPLATE="${mtool}/template";
+#export MTEMPLATE="${mtool}/template";
 
 # linux path to bat utilities folder
 # use m.fgbat() and m.bgbat() to run them
@@ -442,3 +458,17 @@ export MBAT="${mtool}/bat"
 # to access this collection. 
 export BBAT=$(m.winpath "${MBAT}");
 export PATH="${mtool}/bin:${PATH}";
+export PATH="${MBAT}:${PATH}";
+
+if [[  ${JOB##*/} == 'snap'  ]]; then
+export PATH="${JOB}/bin:${PATH}";
+export PATH="${JOB}/bat:${PATH}";
+
+#echo 'snap bin bat is on path';
+#else
+#echo 'not a snap project';
+
+fi
+
+
+source snaplog.sh;
